requirements.txt
opencv-python==4.11.0.86
easyocr==1.7.2
torch==2.7.1
numpy>=1.21.2
Pillow==11.2.1
scikit-image==0.25.2
python-bidi==0.6.6
PyYAML==6.0.2
Shapely==2.1.1
pyclipper==1.3.0.post6
ninja==1.11.1.4
